eal: merge corelist and core mapping options
The "-l" EAL parameter supported a subset of the features that were
provided by the "--lcores" long option. There is no need to have two
different options with different behaviour in this case, so we can just
eliminate the special-case handling for "-l", and have it as a shortened
form of "--lcores".

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>