eal: simplify running CPUs with ids above max lcores
To simplify use on systems with large numbers of cores, where it becomes
more awkward to use CPU cores numbered above 128, we add to EAL the "-L"
or "--lcores-remapped" option, which simply takes a list of cores to run
on, and maps them to lcore ids starting at a defined base (0 by
default).

To enable secondary process use of this, we also add in an extra flag
--lcoreid-base, which changes the starting lcore id from 0. This flag is
mandatory in secondary processes, to ensure that we don't have lcore ids
that overlap between primary and secondary.

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>